refactor: #19 ORDER 전략 Enum화 및 Image와 자잘한 DTO 수정
