+++ src/common/enum/Order.ts
@@ -1,0 +1,4 @@
+export enum Order {
+  ASC = 'ASC',
+  DESC = 'DESC',
+}
