+++ src/entities/image.ts
@@ -32,8 +32,8 @@ export class Image {
   url: string;
 
   @CreateDateColumn()
-  createdAt: string;
+  createdAt: Date;
 
   @UpdateDateColumn()
-  updatedAt: string;
+  updatedAt: Date;
 }
